+++ CHANGES.md
@@ -6,6 +6,8 @@ releases are available on [Anaconda.org](https://anaconda.org/conda-forge/gettsi
 
 ## Unpublished
 
+- {gh}`700` Data columns overwrite functions regardless of time unit
+  ({ghuser}`lars-reimann`).
 - {gh}`694` Replace `tu` groupings with correct groupings. ({ghuser}`MImmesberger`,
   \{ghuser} `lars-reimann`).
 - {gh}`676` Add explicit parent-child links. ({ghuser}`MImmesberger`).